requirements.txt
# ============================================
# Voice IoT Controller - Requirements
# ============================================
# 한국어 음성 명령으로 ESP32 스마트홈 디바이스를 제어하는 시스템
# Python 3.12+ 필요
# ============================================

# ────────────────────────────────────────────
# 웹 프레임워크 & 서버
# ────────────────────────────────────────────
fastapi==0.129.0
uvicorn[standard]==0.41.0
websockets==16.0
starlette==0.52.1

# ────────────────────────────────────────────
# STT (Speech-to-Text) - 음성 인식
# ────────────────────────────────────────────
faster-whisper==1.2.1
openai-whisper
sounddevice==0.5.5
soundfile==0.13.1
numpy==2.4.2
noisereduce==3.0.3

# ────────────────────────────────────────────
# LLM (Large Language Model) - 자연어 처리
# ────────────────────────────────────────────
httpx==0.28.1
ollama==0.6.1

# ────────────────────────────────────────────
# TTS (Text-to-Speech) - 음성 합성
# ────────────────────────────────────────────
edge-tts==7.2.7

# ────────────────────────────────────────────
# 웨이크워드 감지
# ────────────────────────────────────────────
pvporcupine==4.0.2
openwakeword==0.4.0

# ────────────────────────────────────────────
# 설정 & 유틸리티
# ────────────────────────────────────────────
PyYAML==6.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1

# ────────────────────────────────────────────
# HTTP 클라이언트 & WebSocket
# ────────────────────────────────────────────
requests==2.32.5
websocket-client==1.7.0

# ────────────────────────────────────────────
# GUI (선택사항 - PyQt6 대시보드)
# ────────────────────────────────────────────
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
pyqtgraph==0.14.0

# ────────────────────────────────────────────
# TTS - Kokoro (선택사항 - 로컬 ONNX TTS)
# ────────────────────────────────────────────
kokoro-onnx==0.5.0
onnxruntime==1.24.2

# ────────────────────────────────────────────
# 테스트
# ────────────────────────────────────────────
pytest==9.0.2
pytest-asyncio==1.3.0

# ────────────────────────────────────────────
# 기타 의존성
# ────────────────────────────────────────────
typing-extensions==4.15.0
anyio==4.12.1
idna==3.11
certifi==2026.1.4
sniffio==1.3.1
h11==0.16.0
httpcore==1.0.9
charset-normalizer==3.4.4
urllib3==2.6.3

# ============================================
# 설치 방법:
# ============================================
# pip install -r requirements.txt
#
# 또는 선택적 설치:
# pip install -r requirements.txt --no-deps
# pip install fastapi uvicorn faster-whisper edge-tts pvporcupine httpx
# ============================================
